// src/app/Interfaces/ICliente.ts
export interface ICliente {
    cd_cliente?: number;
    nome?: string;
    cpf?: string;
    endereco?: string;
    uf?: string;
    celular?: string;
    tel_fixo?: string;
    complemento?: string;
    estado?: string;
    rg?: string;
    numero?: string;
    cep?: string;
    bairro?: string;
    isShowMore?: boolean;
}